fix: separated auth controller and service, changed auth response to include whole user object
